+++ src/main/java/vectorwing/farmersdelight/common/block/BuddingTomatoBlock.java
@@ -9,7 +9,6 @@ import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.LevelReader;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.BonemealableBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import vectorwing.farmersdelight.common.registry.ModBlocks;
@@ -23,11 +22,6 @@ public class BuddingTomatoBlock extends BuddingBushBlock implements Bonemealable
 	@Override
 	public BlockState getPlant(BlockGetter world, BlockPos pos) {
 		return ModBlocks.BUDDING_TOMATO_CROP.get().defaultBlockState();
-	}
-
-	@Override
-	protected boolean mayPlaceOn(BlockState pState, BlockGetter pLevel, BlockPos pPos) {
-		return pState.is(ModBlocks.RICH_SOIL_FARMLAND.get()) || pState.is(Blocks.FARMLAND);
 	}
 
 	@Override